Moves used functions of sref-utils into the configRouting file and removed the unused sref-utils
