+++ webofneeds/won-owner-webapp/src/main/webapp/app/pages/about.jsx
@@ -11,7 +11,6 @@ import { get, getIn, toAbsoluteURL } from "../utils.js";
 import { attach } from "../cstm-ng-utils.js";
 import { actionCreators } from "../actions/actions.js";
 import { ownerBaseUrl } from "~/config/default.js";
-import * as srefUtils from "../sref-utils.js";
 import * as generalSelectors from "../redux/selectors/general-selectors.js";
 import * as viewSelectors from "../redux/selectors/view-selectors.js";
 import * as accountUtils from "../redux/utils/account-utils.js";
@@ -274,7 +273,6 @@ const questions = [
 class AboutController {
   constructor(/* arguments <- serviceDependencies */) {
     attach(this, serviceDependencies, arguments);
-    Object.assign(this, srefUtils);
 
     window.ab4dbg = this;
